feat: adjust sonner styling and organize types
